requirements.txt
appdirs==1.4.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.8.2
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
et-xmlfile==1.0.1
Flask==3.0.3
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gcovr==5.0
h5py==3.6.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==0.17.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.5.1
mercurial==6.1.1
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
netifaces==0.11.0
networkx==2.4
nltk==3.7
nose==1.3.7
numba==0.55.1
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
python-dateutil==2.8.1
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
regex==2021.11.10
reportlab==3.6.8
requests==2.25.1
roman==3.3
scour==0.38.2
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
SimPy==2.3.1
six==1.16.0
sklearn-pandas==2.0.3
soupsieve==2.3.1
speedtest-cli==2.1.3
ssh-import-id==5.11
sympy==1.9
tables==3.7.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.57.0
Twisted==22.1.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
Werkzeug==3.0.4
xdg==5
xkit==0.0.0
xlwt==1.3.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==5.4.0
